Readme Advanced Deep spaces brief.
Added a Read me protocol for Halo instructor IT academy.  DDAY end of this year. I need all IT entity thats interested in this problem to get into contact with me ASAP . For individual with a criminal record but have advanced engineering skill to apply as well.  Because it comes to my attention that it not the record that defy who you are and where humanity should place you because who your record say you or but it's the individual who defy the status quo a and make the narrative whatever he a she deems it should protect by turning the binary saga around and proofing to the Gods that man can justify audience as being a in a redemption posture. That in the academy we don't look backwards for the past cannot be humanity future. For a mistake is only a impact with will .. To try again and again until the next generation get it right. A man time is not how we got her . Its where we are going next. We challenged are self in this Time and we slow to recognize why we are stagnate in space. Because are mission was never to remain on earth. We are explorers . We build to go up . If we build we don't build we build so it can last. Because we never know when are if we will return. So we build so whoever we leave behind can pick up where we left off . Know being space explorers a couple of privacy we are not privileged to have. Because the technology we need to build the ship we need we need from all the human on earth. Like a onion each pill has its purpose . Just have to find where you belong in all of this. But no matter if your not computer literate it okay with earth mission. And we in the market for life and a sustainable journey to use all . 
To do this a be successful in long run we have to get Earth population properly educated. That's means at least 70 % of the population should be a able to pass the military test for basic training. I say around 2030 that can be achieved.  To be continued my . I believe I've just came up with a way to stop asteroids without having to use two many nuclear war heads on it .  I'm sure it will will work. Conditioning is the best why to prepare for incoming training. Professor Dr Warith Akbar  Advanced engineering Center.  
Seattle